Added a library inspection script
The goal of this script is to instance all classes, run all
functions and methods, get all return values and generate a
report for each attribute. The reports will be at the root
folder under the name of result_lib.import.path.txt (for
example: result_bge.logic.txt). Edit the dict MODULES on
inspect_modules.py to choose which modules to inspect.